# src/main/resources/static/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DataElementHub Rest API",
    "contact": {
      "name": "Medical Informatics Group",
      "url": "http://www.mig-frankfurt.de",
      "email": "vengadeswaran@med.uni-frankfurt.de"
    },
    "description": "REST API Documentation!\n\n  Github repositories:\n\n<https://github.com/mig-frankfurt/dataelementhub.dal>\n\n<https://github.com/mig-frankfurt/dataelementhub.model>\n\n<https://github.com/mig-frankfurt/dataelementhub.rest>\n\n<https://github.com/mig-frankfurt/dataelementhub.gui>\n\nContact:\n\n",
    "version": "1.0.0",
    "license": {
      "url": "https://www.gnu.org/licenses/agpl-3.0.de.html",
      "name": "License: AGPLv3"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/element/{urn}/release": {
      "patch": {
        "tags": [
          "Element"
        ],
        "summary": "Release Element by URN.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
    "/element/{urn}/valuedomain": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get ValueDomain from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/definitions": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Definitions from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/slots": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Slots from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/validation": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Validation from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/identification": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Identification from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/conceptassociation": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get ConceptsAssociations from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
    "/namespaces/{namespaceId}/release": {
      "patch": {
        "tags": [
          "Namespace"
        ],
        "summary": "Release Namespace by id.",
        "operationId": "release_1",
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/namespaces/?scope={scope}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Get all Namespaces with specified scope.",
        "operationId": "",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["read", "write", "admin"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/namespaces/{namespaceId}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Get Namespace by id.",
        "operationId": "read_1",
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Accept-Language",
            "type": "string",
            "required": false,
            "example": "de-DE"
          }

        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Namespace"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Namespace"
        ],
        "summary": "Delete Namespace by id.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/namespaces": {
      "post": {
        "tags": [
          "Namespace"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "NAMESPACE": {
                  "description": "NAMESPACE",
                  "value": {
                    "identification": {
                      "elementType": "NAMESPACE",
                      "hideNamespace": false,
                      "status": "RELEASED"
                    },
                    "definitions": [
                      {
                        "designation": "Namespace designation",
                        "definition": "Namespace definition",
                        "language": "en"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}